requirements.txt
alembic==1.14.1
aniso8601==10.0.0
attrs==25.1.0
blinker==1.8.2
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mailman==1.1.1
Flask-Migrate==4.1.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
googletrans==4.0.0rc1
greenlet==3.1.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.9
MarkupSafe==2.1.5
marshmallow==3.26.1
mysql-connector-python==8.4.0
packaging==24.2
python-dotenv==1.0.1
pytz==2025.1
referencing==0.36.2
requests==2.31.0
rfc3986==1.5.0
rpds-py==0.22.3
sniffio==1.3.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.2.1
Werkzeug==3.0.3
WTForms==3.2.1
